+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Electrum plugin: Restore the previous `oneserver` setting when the plugin is disabled,
+  to prevent users from inadvertently connecting to public Electrum servers with this setting still on.
+
 ## 0.1.1 - 2020-05-27
 
 - Make bwt available as an Electrum plugin! 💥